requirements.txt
python-telegram-bot
openai-whisper
torch
ffmpeg-python
python-dotenv
